services/knowledge/pyproject.toml
[tool.poetry]
name = "hill90-knowledge"
version = "0.1.0"
description = "Hill90 Agent Knowledge Manager (AKM) — persistent knowledge store for agents"
authors = ["Hill90 Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
asyncpg = "^0.30.0"
pyyaml = "^6.0"
PyJWT = {extras = ["crypto"], version = "^2.8.0"}
cryptography = "^44.0.0"
structlog = "^24.1.0"
opentelemetry-distro = ">=0.60b1"
opentelemetry-exporter-otlp-proto-http = "^1.39.1"
opentelemetry-instrumentation-fastapi = ">=0.60b1"
opentelemetry-instrumentation-asyncpg = ">=0.60b1"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
ruff = "^0.1.0"
mypy = "^1.8.0"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m not integration')",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
